refactor(app): decompose SpecTabs into feature folder
- Split the 961-line SpecTabs.tsx monolith into src/sections/spec-detail/SpecTabs/
  with index.tsx (tabs orchestration, tag-count state, tag chip rendering),
  CodeTab.tsx, SpecTab.tsx, ImplTab.tsx, QualityTab.tsx, and utils.tsx
- utils.tsx keeps the module-level tag-count cache (via getter/setter to preserve
  its once-per-session lifetime), Md helpers, parseInlineCode, TabPanel,
  formatDate, and the tag→URL-param maps
- QualityTab expansion state stays lifted in index.tsx (as before) and is passed
  down as expandedCategories/onToggleCategory props; copy state moved with the
  copy button into CodeTab
- No behavior change: all trackEvent calls, aria attributes, and interactions
  preserved verbatim; existing SpecTabs.test.tsx moved into the folder unchanged
- Add focused per-tab tests for the now prop-driven surfaces (QualityTab
  expansion contract, CodeTab copy/tracking, ImplTab date fallback, SpecTab
  inline-code rendering): 60→64 test files, 552→572 tests

Part 9 of the frontend modernization roadmap.
